pyproject.toml
[project]
name = "CSE472-Project1"
version = "0.1.0"
description = "Repository for Project 1 of CSE472: Social Media Mining (Fall 2025)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mastodon-py>=2.1.3",
]
